1.Переместил папку screens в ui 2.Сделал макет фрагмента для получения адресса польлзователя
