--- app/src/main/java/com/example/wsrpractice/presentetion/screens/checkout_order/CheckoutOrderViewModel.kt
@@ -1,9 +1,0 @@
-package com.example.wsrpractice.presentetion.screens.checkout_order
-
-import androidx.lifecycle.ViewModel
-
-class CheckoutOrderViewModel: ViewModel() {
-
-
-
-}
