+++ app/src/main/java/com/example/wsrpractice/presentetion/ui/activity/MainActivity.kt
@@ -6,13 +6,14 @@ import androidx.core.view.isVisible
 import com.example.wsrpractice.App
 import com.example.wsrpractice.R
 import com.example.wsrpractice.databinding.ActivityMainBinding
-import com.example.wsrpractice.presentetion.screens.Screens
-import com.example.wsrpractice.presentetion.screens.analyze.FragmentAnalyze
-import com.example.wsrpractice.presentetion.screens.analyze.bottom_sheet.FragmentBottomSheetDialogDetailAnalyze
-import com.example.wsrpractice.presentetion.screens.checkout_order.FragmentCheckoutOrder
-import com.example.wsrpractice.presentetion.screens.nothing.FragmentNothing
-import com.example.wsrpractice.presentetion.screens.profile.FragmentProfile
-import com.example.wsrpractice.presentetion.ui.fragment.*
+import com.example.wsrpractice.databinding.FragmentSetUserAddressBinding
+import com.example.wsrpractice.presentetion.ui.screens.Screens
+import com.example.wsrpractice.presentetion.ui.screens.analyze.FragmentAnalyze
+import com.example.wsrpractice.presentetion.ui.screens.analyze.bottom_sheet.FragmentBottomSheetDialogDetailAnalyze
+import com.example.wsrpractice.presentetion.ui.screens.checkout_order.main.FragmentCheckoutOrder
+import com.example.wsrpractice.presentetion.ui.screens.checkout_order.map.FragmentSetUserAddress
+import com.example.wsrpractice.presentetion.ui.screens.nothing.FragmentNothing
+import com.example.wsrpractice.presentetion.ui.screens.profile.FragmentProfile
 import com.github.terrakok.cicerone.androidx.AppNavigator
 
 class MainActivity : AppCompatActivity() {
@@ -30,7 +31,7 @@ class MainActivity : AppCompatActivity() {
 //        router.replaceScreen(Screens.analyze())
 
         supportFragmentManager.beginTransaction()
-            .replace(R.id.fragmentContainerView,FragmentCheckoutOrder())
+            .replace(R.id.fragmentContainerView, FragmentSetUserAddress())
             .commit()
         initBottomNav()
