+++ app/src/main/java/com/example/wsrpractice/presentetion/ui/screens/checkout_order/CheckoutOrderViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.wsrpractice.presentetion.ui.screens.checkout_order
+
+import androidx.lifecycle.ViewModel
+
+class CheckoutOrderViewModel: ViewModel() {
+
+
+
+}
