+++ app/src/main/java/com/example/wsrpractice/presentetion/ui/screens/nothing/FragmentNothing.kt
@@ -1,0 +1,17 @@
+package com.example.wsrpractice.presentetion.ui.screens.nothing
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.wsrpractice.core.ui.BaseFragment
+import com.example.wsrpractice.databinding.FragmentNothingScreenBinding
+
+class FragmentNothing:BaseFragment<FragmentNothingScreenBinding>(
+    FragmentNothingScreenBinding::inflate
+) {
+
+
+
+}
