+++ app/src/main/res/drawable/bottom_sheet_line.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="91dp"
+    android:height="2dp"
+    android:viewportWidth="91"
+    android:viewportHeight="2">
+  <path
+      android:pathData="M1,1H90"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#636262"
+      android:strokeLineCap="round"/>
+</vector>
