+++ app/src/main/res/drawable/ic_my_location.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="55dp"
+    android:height="55dp"
+    android:viewportWidth="55"
+    android:viewportHeight="55">
+  <path
+      android:pathData="M27.5,0L27.5,0A27.5,27.5 0,0 1,55 27.5L55,27.5A27.5,27.5 0,0 1,27.5 55L27.5,55A27.5,27.5 0,0 1,0 27.5L0,27.5A27.5,27.5 0,0 1,27.5 0z"
+      android:fillColor="#1A6FEE"/>
+  <path
+      android:pathData="M27.5,20.409C23.582,20.409 20.409,23.582 20.409,27.5C20.409,31.418 23.582,34.591 27.5,34.591C31.418,34.591 34.591,31.418 34.591,27.5C34.591,23.582 31.418,20.409 27.5,20.409ZM43.348,25.727C42.533,18.335 36.665,12.467 29.273,11.652V8H25.727V11.652C18.335,12.467 12.467,18.335 11.652,25.727H8V29.273H11.652C12.467,36.665 18.335,42.533 25.727,43.348V47H29.273V43.348C36.665,42.533 42.533,36.665 43.348,29.273H47V25.727H43.348ZM27.5,39.909C20.639,39.909 15.091,34.361 15.091,27.5C15.091,20.639 20.639,15.091 27.5,15.091C34.361,15.091 39.909,20.639 39.909,27.5C39.909,34.361 34.361,39.909 27.5,39.909Z"
+      android:fillColor="#ffffff"/>
+</vector>
