+++ app/src/main/res/layout/fragment_basket.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentetion.screens.basket.FragmentBasket"
+    tools:context=".presentetion.ui.screens.basket.FragmentBasket"
     android:orientation="vertical"
     android:paddingHorizontal="20dp"
     android:paddingVertical="32dp">
